feat: implement real time update server on top of web sockets
Closes: https://github.com/talos-systems/theila/issues/2

Add backend server that is able to connect to Kubernetes `controller-runtime`
and fetch data from it and send updates to the JS frontend through a web socket.

As we have pretty similar implementation of `controller-runtime` in Talos all
communication with K8s runtime is done through an abstraction layer.

Potentially, we should get exactly the same interface for Talos resource
watch.

WebSocket protocol is currently used for all client-server
communication, but later on we may decide to add REST RPC calls as well.
Mainly for Talos API calls.

UI is still more of a draft, but it is possible to get updated list of
pods, nodes, namespaces, deployments, daemonsets.

Signed-off-by: Artem Chernyshev <artem.0xD2@gmail.com>