feat: implement nodes view
Fixes: https://github.com/talos-systems/theila/issues/15

Nodes view is currently pulling `nodes.v1` resource from the target
cluster, which is going to be replaced by `machines` from the bootstrap
cluster with a label selector later on.

Introduced a new feature into Watch API that allows client to specify
the context to use for the watch.

Now `WatchSpec` has `Context` message inside. Then it can be either:
- get the context config from the Kubeconfig.
- get the Kubeconfig from the current cluster using Kubernetes secret
that is created by Sidero. Which normally has the following name
`<cluster-name>-kubeconfig`.

If the second flow is used, Kubernetes runtime creates the config
context cache. Cluster UID is used to identify the context name, to
avoid conflicts when the cluster is deleted and recreated with the same
name.

Runtime interface now has two more methods:
- `AddContext` which can help to register new contexts in the runtime.
- `Get` which implements both list and get APIs. Then it can be used as
part of gRPC API gateway methods to allow client do single time requests
for the data (apart from watch ones).

Also added a preliminary version of Talos runtime that can
only return Kubeconfig for now, given that it is provided the correct
Talosconfig with endpoints and nodes.

Signed-off-by: Artem Chernyshev <artem.0xD2@gmail.com>