feat: detect cluster capabilities
From now on Theila supports clusters without Sidero installed.
When it starts it checks the registered CRDs to see if there are CAPI,
Sidero and Packet (even if we are not using this one yet) CRDs registered.

This change required to alter our URI scheme to make `cluster`, `uid`
and `namespace` resources optional as views for nodes, pods are now
populated at the current cluster level as well.

Signed-off-by: Artem Chernyshev <artem.0xD2@gmail.com>