refactor: use Metadata to pass context in all gRPC calls
Remove context name and CAPI cluster params from gRPC requests
everywhere and move it to the metadata.
The same as it was done for `Logs` call which is being proxied to Talos
through gRPC proxy.

This will align Theila internal methods with proxied ones.

Additionally make Theila listen on the [::] address instead of `0.0.0.0`
to harden security at least a bit.

Signed-off-by: Artem Chernyshev <artem.0xD2@gmail.com>