feat: rework the node overview page
Show detailed node information instead of charts.
That comes in scope of global adjustment of styles to properly handle
different screen size, as it has become clear that it's impossible to
show long values in the table view right now.

Merge overview page and services page.
Move charts and processes to the Monitoring page.

Additionally remove wrong `md:` definitions in many places and properly
handle breakpoints on the ultra-wide screens.

Signed-off-by: Artem Chernyshev <artem.0xD2@gmail.com>