chore: allow getting resources without version and group
That should help to keep Theila backward and forward compatible with
any CRDs from CAPI or whatever else.

Now if the resource name is represented only with a single word without
any dots in it, the backend will look up version and group by using the
discovery service. Same way as it is done by kubectl.

Signed-off-by: Artem Chernyshev <artem.chernyshev@talos-systems.com>